ref and add dequeue if
