+++ zoo/src/main/scala/com/intel/analytics/zoo/serving/engine/FlinkRedisSink.scala
@@ -18,7 +18,7 @@
 package com.intel.analytics.zoo.serving.engine
 
 import com.intel.analytics.zoo.serving.ClusterServing
-import com.intel.analytics.zoo.serving.pipeline.RedisIO
+import com.intel.analytics.zoo.serving.pipeline.RedisUtils
 import com.intel.analytics.zoo.serving.utils.{ClusterServingHelper, Conventions}
 import org.apache.flink.configuration.Configuration
 import org.apache.flink.streaming.api.functions.sink.{RichSinkFunction, SinkFunction}
@@ -34,28 +34,8 @@ class FlinkRedisSink(params: ClusterServingHelper)
   override def open(parameters: Configuration): Unit = {
     logger = Logger.getLogger(getClass)
 
-    if (params.redisSecureEnabled) {
-      System.setProperty("javax.net.ssl.trustStore", params.redisSecureTrustStorePath)
-      System.setProperty("javax.net.ssl.trustStorePassword", params.redisSecureTrustStoreToken)
-      System.setProperty("javax.net.ssl.keyStoreType", "JKS")
-      System.setProperty("javax.net.ssl.keyStore", params.redisSecureTrustStorePath)
-      System.setProperty("javax.net.ssl.keyStorePassword", params.redisSecureTrustStoreToken)
-    }
-    if (ClusterServing.jedisPool == null) {
-      ClusterServing.synchronized {
-        if (ClusterServing.jedisPool == null) {
-          ClusterServing.jedisPool = new JedisPool(ClusterServing.jedisPoolConfig,
-            params.redisHost, params.redisPort, params.redisTimeout, params.redisSecureEnabled)
-        }
-      }
-    }
-
-    params.redisSecureEnabled match {
-      case true => logger.info(s"FlinkRedisSink connect to secured Redis successfully.")
-      case false => logger.info(s"FlinkRedisSink connect to plain Redis successfully.")
-    }
-    jedis = RedisIO.getRedisClient(ClusterServing.jedisPool)
-
+    RedisUtils.initializeRedis()
+    jedis = RedisUtils.getRedisClient(ClusterServing.jedisPool)
   }
 
   override def close(): Unit = {
@@ -68,7 +48,7 @@ class FlinkRedisSink(params: ClusterServingHelper)
     val ppl = jedis.pipelined()
     var cnt = 0
     value.foreach(v => {
-      RedisIO.writeHashMap(ppl, v._1, v._2, params.jobName)
+      RedisUtils.writeHashMap(ppl, v._1, v._2, params.jobName)
       if (v._2 != "NaN") {
         cnt += 1
       }
